requirements.txt
langchain==0.0.302
numpy==1.26.4
openai==1.14.1
pandas==2.2.3
PyPDF2==3.0.1
selenium==4.29.0
streamlit==1.42.2
streamlit-extras==0.5.5
streamlit-option-menu==0.4.0
tiktoken==0.5.2
pdf2image==1.17.0
pdfplumber==0.11.4
pytesseract==0.3.13
python-dotenv==1.0.1
google-generativeai
fpdf
